server/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.2
Pillow==8.3.1
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
